Admin only
-Removed the feature to change other's gamemode..
-Removed the feature to change everyone's gamemode [/gmsp -a].
-Made it admin only (due to request).
-Removed console compatibility.
