--- src/me/StevenLawson/TotalFreedomMod/Commands/Command_spectate.java
@@ -1,72 +1,0 @@
-package me.StevenLawson.TotalFreedomMod.Commands;
-
-import me.StevenLawson.TotalFreedomMod.TFM_AdminList;
-import me.StevenLawson.TotalFreedomMod.TFM_Util;
-import org.bukkit.GameMode;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-
-@CommandPermissions(level = AdminLevel.ALL, source = SourceType.ONLY_IN_GAME)
-@CommandParameters(description = "Quickly change your own gamemode to spectate, or define someone's username to change theirs.", usage = "/<command> [partialname]", aliases = "gmsp")
-public class Command_spectate extends TFM_Command
-{
-    @Override
-    public boolean run(CommandSender sender, Player sender_p, Command cmd, String commandLabel, String[] args, boolean senderIsConsole)
-    {
-        if (senderIsConsole)
-        {
-            if (args.length == 0)
-            {
-                sender.sendMessage("You cannot send from console!");
-                return true;
-            }
-        }
-
-        Player player;
-        if (args.length == 0)
-        {
-            player = sender_p;
-        }
-        else
-        {
-            if (args[0].equalsIgnoreCase("-a"))
-            {
-                if (!TFM_AdminList.isSuperAdmin(sender))
-                {
-                    sender.sendMessage(TFM_Command.MSG_NO_PERMS);
-                    return true;
-                }
-
-                for (Player targetPlayer : server.getOnlinePlayers())
-                {
-                    targetPlayer.setGameMode(GameMode.SPECTATOR);
-                }
-
-                TFM_Util.adminAction(sender.getName(), "Changing everyone's gamemode to spectate", false);
-                return true;
-            }
-
-            if (!(senderIsConsole || TFM_AdminList.isSuperAdmin(sender)))
-            {
-                playerMsg("Only superadmins can change other user's gamemode.");
-                return true;
-            }
-
-            player = getPlayer(args[0]);
-
-            if (player == null)
-            {
-                sender.sendMessage(TFM_Command.PLAYER_NOT_FOUND);
-                return true;
-            }
-
-        }
-
-        playerMsg("Setting " + player.getName() + " to game mode 'Spectate'.");
-        playerMsg(player, sender.getName() + " set your game mode to 'Spectate'.");
-        player.setGameMode(GameMode.SPECTATOR);
-
-        return true;
-    }
-}
